fix the memory leak when client sets maxResults as Integer.MAX_VALUE (#30)
When a client specifies the number of results to return in the request, if it sets the value to Integer.MAX_VALUE or it does not set the value, GraphJet library will try to initialize the priority queue to size Integer.MAX_VALUE. This causes JVM to throw an exception and kill the service. This rb tries to fix the issue.